pyproject.toml
[project]
name = "inspertorchaudio"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Fabio Ayres", email = "fabioja@insper.edu.br" }
]
requires-python = ">=3.12"
dependencies = [
    "pandas>=2.3.1",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
    "resampy>=0.4.3",
    "scikit-learn>=1.7.1",
    "soundfile>=0.13.1",
    "toml>=0.10.2",
    "torch>=2.7.1",
    "torchaudio>=2.7.1",
    "tqdm>=4.67.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "ipykernel>=6.30.0",
    "ipython>=9.4.0",
    "jupyter>=1.1.1",
    "pytest>=8.4.1",
    "ruff>=0.12.5",
]

[tool.setuptools.package-data]
"inspertorchaudio" = ["resources/*"]

[tool.ruff]
line-length = 90

[tool.ruff.format]
quote-style = "single"
